+++ helpers/fshelper/globwalkfs.go
@@ -8,7 +8,8 @@ import (
 )
 
 //  GlobWalkFS create a FS that limits the WalkDir function to the
-//  list of files that match the glob expression
+//  list of files that match the glob expression, and cheats to
+//  matches *.XMP files in all circumstances
 //
 //  It implements ReadDir and Stat to filter the file list
 //
@@ -32,7 +33,7 @@ func NewGlobWalkFS(fsys fs.FS, pattern string) (fs.FS, error) {
 // match the current file name with the pattern
 // matches files having a path starting by the patten
 //
-//	ex:  file /path/to/file matches with the pattern */to
+//	ex:  file /path/to/file matches with the pattern /*/to
 func (gw GlobWalkFS) match(name string) (bool, error) {
 	if name == "." {
 		return true, nil
@@ -49,29 +50,15 @@ func (gw GlobWalkFS) match(name string) (bool, error) {
 
 // Open the name only if the name matches with the pattern
 func (gw GlobWalkFS) Open(name string) (fs.File, error) {
-	match, err := gw.match(name)
-	if err != nil {
-		return nil, err
-	}
-	if !match {
-		return nil, fs.ErrNotExist
-	}
 	return gw.rootFS.Open(name)
 }
 
 // Stat the name only if the name matches with the pattern
 func (gw GlobWalkFS) Stat(name string) (fs.FileInfo, error) {
-	match, err := gw.match(name)
-	if err != nil {
-		return nil, err
-	}
-	if !match {
-		return nil, fs.ErrNotExist
-	}
 	return fs.Stat(gw.rootFS, name)
 }
 
-// ReadDir return all DirEntries that match with the pattern
+// ReadDir return all DirEntries that match with the pattern or .XMP files
 func (gw GlobWalkFS) ReadDir(name string) ([]fs.DirEntry, error) {
 	match, err := gw.match(name)
 	if err != nil {
@@ -88,6 +75,15 @@ func (gw GlobWalkFS) ReadDir(name string) ([]fs.DirEntry, error) {
 	returned := []fs.DirEntry{}
 	for _, e := range entries {
 		p := path.Join(name, e.Name())
+
+		// Always matches .XMP files...
+		if !e.IsDir() {
+			ext := strings.ToUpper(path.Ext(e.Name()))
+			if ext == ".XMP" {
+				returned = append(returned, e)
+				continue
+			}
+		}
 		match, _ = gw.match(p)
 		if match {
 			returned = append(returned, e)
@@ -99,9 +95,7 @@ func (gw GlobWalkFS) ReadDir(name string) ([]fs.DirEntry, error) {
 // FixedPathAndMagic split the path with the fixed part and the variable part
 func FixedPathAndMagic(name string) (string, string) {
 	if !HasMagic(name) {
-		d, f := path.Split(name)
-		d = strings.TrimSuffix(d, "/")
-		return d, f
+		return name, ""
 	}
 	name = filepath.ToSlash(name)
 	parts := strings.Split(name, "/")
@@ -111,5 +105,9 @@ func FixedPathAndMagic(name string) (string, string) {
 			break
 		}
 	}
-	return path.Join(parts[:p]...), path.Join(parts[p:]...)
+	fixed := ""
+	if name[0] == '/' {
+		fixed = "/"
+	}
+	return fixed + path.Join(parts[:p]...), path.Join(parts[p:]...)
 }