+++ cmd/duplicate/duplicate.go
@@ -21,9 +21,10 @@ import (
 
 type DuplicateCmd struct {
 	*cmd.SharedFlags
-	AssumeYes      bool             // When true, doesn't ask to the user
-	DateRange      immich.DateRange // Set capture date range
-	IgnoreTZErrors bool             // Enable TZ error tolerance
+	AssumeYes       bool             // When true, doesn't ask to the user
+	DateRange       immich.DateRange // Set capture date range
+	IgnoreTZErrors  bool             // Enable TZ error tolerance
+	IgnoreExtension bool             // Ignore file extensions when checking for duplicates
 
 	assetsByID          map[string]*immich.Asset
 	assetsByBaseAndDate map[duplicateKey][]*immich.Asset
@@ -32,6 +33,7 @@ type DuplicateCmd struct {
 type duplicateKey struct {
 	Date time.Time
 	Name string
+	Type string
 }
 
 func NewDuplicateCmd(ctx context.Context, common *cmd.SharedFlags, args []string) (*DuplicateCmd, error) {
@@ -50,6 +52,7 @@ func NewDuplicateCmd(ctx context.Context, common *cmd.SharedFlags, args []string
 	cmd.BoolFunc("ignore-tz-errors", "Ignore timezone difference to check duplicates (default: FALSE).", myflag.BoolFlagFn(&app.IgnoreTZErrors, false))
 	cmd.BoolFunc("yes", "When true, assume Yes to all actions", myflag.BoolFlagFn(&app.AssumeYes, false))
 	cmd.Var(&app.DateRange, "date", "Process only documents having a capture date in that range.")
+	cmd.BoolFunc("ignore-extension", "When true, ignores extensions when checking for duplicates (default: FALSE)", myflag.BoolFlagFn(&app.IgnoreExtension, false))
 	err := cmd.Parse(args)
 	if err != nil {
 		return nil, err
@@ -84,8 +87,12 @@ func DuplicateCommand(ctx context.Context, common *cmd.SharedFlags, args []strin
 		k := duplicateKey{
 			Date: d,
 			Name: strings.ToUpper(a.OriginalFileName + path.Ext(a.OriginalPath)),
+			Type: a.Type,
 		}
 
+		if app.IgnoreExtension {
+			k.Name = strings.TrimSuffix(k.Name, path.Ext(a.OriginalPath))
+		}
 		l := app.assetsByBaseAndDate[k]
 		if len(l) > 0 {
 			dupCount++