+++ helpers/configuration/doc.go
@@ -1,0 +1,4 @@
+/*
+This package handles the configuration file stored into user's machine profile
+*/
+package configuration